[feat] Re_Day_4 완료
1. 공통응답 처리 및 예외처리 완료
