--- src/main/java/org/example/springstudy/auth/controller/dto/LoginRequest.java
@@ -1,7 +1,0 @@
-package org.example.springstudy.auth.controller.dto;
-
-public record LoginRequest(
-    String username,
-    String password
-) {
-}
