--- src/main/java/org/example/springstudy/auth/controller/dto/LoginResponse.java
@@ -1,12 +1,0 @@
-package org.example.springstudy.auth.controller.dto;
-
-public record LoginResponse(
-    String accessToken,
-    String tokenType
-) {
-    private final static String TOKEN_TYPE = "bearer ";
-
-    public static LoginResponse from(String accessToken) {
-        return new LoginResponse(accessToken, TOKEN_TYPE);
-    }
-}
