+++ src/main/java/org/example/springstudy/domain/auth/controller/dto/LoginRequest.java
@@ -1,0 +1,7 @@
+package org.example.springstudy.domain.auth.controller.dto;
+
+public record LoginRequest(
+    String username,
+    String password
+) {
+}
