+++ src/main/java/org/example/springstudy/domain/auth/controller/dto/LoginResponse.java
@@ -1,0 +1,12 @@
+package org.example.springstudy.domain.auth.controller.dto;
+
+public record LoginResponse(
+    String accessToken,
+    String tokenType
+) {
+    private final static String TOKEN_TYPE = "bearer ";
+
+    public static LoginResponse from(String accessToken) {
+        return new LoginResponse(accessToken, TOKEN_TYPE);
+    }
+}
