+++ src/main/java/org/example/springstudy/domain/hello/service/HelloService.java
@@ -1,0 +1,11 @@
+package org.example.springstudy.domain.hello.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class HelloService {
+
+    public String hi() {
+        return "hello";
+    }
+}
