+++ src/main/java/org/example/springstudy/domain/member/controller/dto/MemberResponse.java
@@ -1,0 +1,13 @@
+package org.example.springstudy.domain.member.controller.dto;
+
+import org.example.springstudy.domain.member.entity.MemberEntity;
+
+public record MemberResponse(
+    Long id,
+    String username,
+    String nickname
+) {
+    public static MemberResponse from(MemberEntity member) {
+        return new MemberResponse(member.getId(), member.getUsername(), member.getNickname());
+    }
+}
