--- src/main/java/org/example/springstudy/hello/controller/HelloController.java
@@ -1,20 +1,0 @@
-package org.example.springstudy.hello.controller;
-
-import org.example.springstudy.hello.service.HelloService;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloController {
-
-    private final HelloService helloService;
-
-    public HelloController(HelloService helloService) {
-        this.helloService = helloService;
-    }
-
-    @GetMapping("/hello")
-    public String hello() {
-        return helloService.hi();
-    }
-}
