--- src/main/java/org/example/springstudy/hello/service/HelloService.java
@@ -1,11 +1,0 @@
-package org.example.springstudy.hello.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class HelloService {
-
-    public String hi() {
-        return "hello";
-    }
-}
