--- src/main/java/org/example/springstudy/post/controller/PostController.java
@@ -1,54 +1,0 @@
-package org.example.springstudy.post.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.example.springstudy.post.controller.dto.PostRequest;
-import org.example.springstudy.post.controller.dto.PostResponse;
-import org.example.springstudy.post.service.PostService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/posts")
-@RequiredArgsConstructor
-public class PostController {
-    private final PostService postService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public PostResponse createPost(@RequestBody PostRequest request) {
-        return postService.save(request);
-    }
-
-    @GetMapping
-    @ResponseStatus(HttpStatus.OK)
-    public List<PostResponse> getAllPosts() {
-        return postService.getAllPosts();
-    }
-
-    @GetMapping("/{postId}")
-    public ResponseEntity<PostResponse> getPostById(@PathVariable Long postId) {
-        try {
-            return ResponseEntity.ok(postService.findById(postId));
-        } catch (RuntimeException e) {
-            return ResponseEntity.notFound().build();
-        }
-    }
-
-    @PutMapping("/{postId}")
-    public ResponseEntity<PostResponse> updatePost(@PathVariable Long postId, @RequestBody PostRequest request) {
-        try {
-            return ResponseEntity.ok(postService.update(request, postId));
-        } catch (RuntimeException e) {
-            return ResponseEntity.notFound().build();
-        }
-    }
-
-    @DeleteMapping("/{postId}")
-    public ResponseEntity<PostResponse> deletePost(@PathVariable Long postId) {
-        postService.deleteById(postId);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-}
