--- src/main/java/org/example/springstudy/post/entity/PostEntity.java
@@ -1,25 +1,0 @@
-package org.example.springstudy.post.entity;
-
-import lombok.Getter;
-
-@Getter
-public class PostEntity {
-    private static long AUTO_INCREMENT_ID = 1;
-
-    private Long id;
-    private String title;
-    private String content;
-    private String author;
-
-    public PostEntity(String title, String content, String author) {
-        this.id = AUTO_INCREMENT_ID++;
-        this.title = title;
-        this.content = content;
-        this.author = author;
-    }
-
-    public void update(String title, String content) {
-        this.title = title;
-        this.content = content;
-    }
-}
